+++ packages/abuse/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xnetjs/abuse
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/identity@0.10.0
+  - @xnetjs/crypto@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes