+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @xnetjs/cli
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`0721fd5`](https://github.com/crs48/xNet/commit/0721fd5d263abd3242a3b10cf827fa552cbacbb7)]:
+  - @xnetjs/data@0.10.0
+  - @xnetjs/plugins@0.10.0
+  - @xnetjs/runtime@0.2.2
+  - @xnetjs/sqlite@0.10.0
+  - @xnetjs/sync@0.10.0
+  - @xnetjs/identity@0.10.0
+  - @xnetjs/crypto@0.10.0
+  - @xnetjs/core@0.10.0
+
 ## 0.1.1
 
 ### Patch Changes