+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xnetjs/crypto
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/core@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes