+++ packages/identity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @xnetjs/identity
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/crypto@0.10.0
+  - @xnetjs/core@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes